Resolve clang warnings about wrong format specifiers for NSLog. Patch by Sebastian Reitenbach <sebastia@l00-bugdead-prods.de>
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@36133 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSApplication.m
@@ -2091,16 +2091,16 @@ See -runModalForWindow:
 
       default:	/* pass all other events to the event's window	*/
 	{
-	  NSWindow	*window = [theEvent window];
+	  NSWindow *window = [theEvent window];
 
 	  if (!theEvent)
 	    NSDebugLLog(@"NSEvent", @"NSEvent is nil!\n");
 	  if (type == NSMouseMoved)
 	    NSDebugLLog(@"NSMotionEvent", @"Send move (%d) to %@", 
-			type, window);
+			(int)type, window);
 	  else
 	    NSDebugLLog(@"NSEvent", @"Send NSEvent type: %d to %@", 
-			type, window);
+			(int)type, window);
 	  if (window)
 	    [window sendEvent: theEvent];
 	  else if (type == NSRightMouseDown)
